Update to latest from Campsych master

+++ src/Concerto/PanelBundle/Controller/DataTableController.php
@@ -3,7 +3,6 @@
 namespace Concerto\PanelBundle\Controller;
 
 use Concerto\PanelBundle\Service\FileService;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 use Symfony\Component\HttpFoundation\StreamedResponse;
@@ -11,7 +10,6 @@ use Concerto\PanelBundle\DAO\DAOUnsupportedOperationException;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Translation\TranslatorInterface;
 use Concerto\PanelBundle\Service\DataTableService;
 use Concerto\PanelBundle\Service\ImportService;
@@ -27,13 +25,13 @@ class DataTableController extends AExportableTabController
 {
 
     const ENTITY_NAME = "DataTable";
-    const EXPORT_FILE_PREFIX = "DataTable_";
+    const EXPORT_FILE_PREFIX = "DataTable";
 
     private $userService;
 
-    public function __construct($environment, EngineInterface $templating, DataTableService $service, TranslatorInterface $translator, TokenStorageInterface $securityTokenStorage, ImportService $importService, ExportService $exportService, UserService $userService, FileService $fileService)
+    public function __construct($environment, EngineInterface $templating, DataTableService $service, TranslatorInterface $translator, ImportService $importService, ExportService $exportService, UserService $userService, FileService $fileService)
     {
-        parent::__construct($environment, $templating, $service, $translator, $securityTokenStorage, $importService, $exportService, $fileService);
+        parent::__construct($environment, $templating, $service, $translator, $importService, $exportService, $fileService);
 
         $this->entityName = self::ENTITY_NAME;
         $this->exportFilePrefix = self::EXPORT_FILE_PREFIX;
@@ -63,6 +61,17 @@ class DataTableController extends AExportableTabController
     }
 
     /**
+     * @Route("/DataTable/{object_id}/toggleLock", name="DataTable_toggleLock")
+     * @param Request $request
+     * @param $object_id
+     * @return Response
+     */
+    public function toggleLock(Request $request, $object_id)
+    {
+        return parent::toggleLock($request, $object_id);
+    }
+
+    /**
      * @Route("/DataTable/form/{action}", name="DataTable_form", defaults={"action":"edit"})
      * @param string $action
      * @param array $params
@@ -74,30 +83,33 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{object_id}/save", name="DataTable_save")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{object_id}/save", name="DataTable_save", methods={"POST"})
      * @param Request $request
      * @param $object_id
      * @return Response
      */
     public function saveAction(Request $request, $object_id)
     {
+        if (!$this->service->canBeModified($object_id, $request->get("updatedOn"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         $result = $this->service->save(
-            $this->securityTokenStorage->getToken()->getUser(), //
-            $object_id, //
-            $request->get("name"), //
-            $request->get("description"), //
-            $request->get("accessibility"), //
-            $request->get("archived") === "1", //
-            $this->userService->get($request->get("owner")), //
-            $request->get("groups") //
+            $object_id,
+            $request->get("name"),
+            $request->get("description"),
+            $request->get("accessibility"),
+            $request->get("archived") === "1",
+            $this->userService->get($request->get("owner")),
+            $request->get("groups")
         );
         return $this->getSaveResponse($result);
     }
 
     /**
-     * @Route("/DataTable/{object_id}/copy", name="DataTable_copy")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{object_id}/copy", name="DataTable_copy", methods={"POST"})
      * @param Request $request
      * @param $object_id
      * @return Response
@@ -108,14 +120,14 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{object_ids}/delete", name="DataTable_delete")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{object_ids}/delete", name="DataTable_delete", methods={"POST"})
+     * @param Request $request
      * @param string $object_ids
      * @return Response
      */
-    public function deleteAction($object_ids)
+    public function deleteAction(Request $request, $object_ids)
     {
-        return parent::deleteAction($object_ids);
+        return parent::deleteAction($request, $object_ids);
     }
 
     /**
@@ -216,14 +228,20 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{table_id}/column/{column_names}/delete", name="DataTable_column_delete")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/column/{column_names}/delete", name="DataTable_column_delete", methods={"POST"})
+     * @param Request $request
      * @param $table_id
      * @param string $column_names
      * @return Response
      */
-    public function deleteColumnAction($table_id, $column_names)
+    public function deleteColumnAction(Request $request, $table_id, $column_names)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         $this->service->deleteColumns($table_id, $column_names);
         $response = new Response(json_encode(array("result" => 0)));
         $response->headers->set('Content-Type', 'application/json');
@@ -231,14 +249,20 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{table_id}/row/{row_ids}/delete", name="DataTable_row_delete")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/row/{row_ids}/delete", name="DataTable_row_delete", methods={"POST"})
+     * @param Request $request
      * @param $table_id
      * @param string $row_ids
      * @return Response
      */
-    public function deleteRowAction($table_id, $row_ids)
+    public function deleteRowAction(Request $request, $table_id, $row_ids)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         $this->service->deleteRows($table_id, $row_ids);
         $response = new Response(json_encode(array("result" => 0)));
         $response->headers->set('Content-Type', 'application/json');
@@ -246,13 +270,19 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{table_id}/truncate", name="DataTable_truncate")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/truncate", name="DataTable_truncate", methods={"POST"})
+     * @param Request $request
      * @param $table_id
      * @return Response
      */
-    public function truncateAction($table_id)
+    public function truncateAction(Request $request, $table_id)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         $this->service->truncate($table_id);
         $response = new Response(json_encode(array("result" => 0)));
         $response->headers->set('Content-Type', 'application/json');
@@ -260,8 +290,27 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{table_id}/column/{column_name}/save", name="DataTable_column_save")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/deleteAll", name="DataTable_deleteAll", methods={"POST"})
+     * @param Request $request
+     * @param $table_id
+     * @return Response
+     */
+    public function deleteAllAction(Request $request, $table_id)
+    {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
+        $this->service->deleteAll($table_id);
+        $response = new Response(json_encode(array("result" => 0)));
+        $response->headers->set('Content-Type', 'application/json');
+        return $response;
+    }
+
+    /**
+     * @Route("/DataTable/{table_id}/column/{column_name}/save", name="DataTable_column_save", methods={"POST"})
      * @param Request $request
      * @param $table_id
      * @param string $column_name
@@ -269,31 +318,37 @@ class DataTableController extends AExportableTabController
      */
     public function saveColumnAction(Request $request, $table_id, $column_name)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         try {
             $errors = $this->service->saveColumn($table_id, $column_name, $request->get("name"), $request->get("type"));
-            if (count($errors) > 0) {
-                for ($i = 0; $i < count($errors); $i++) {
-                    $errors[$i] = $this->translator->trans($errors[$i]);
-                }
-                $response = new Response(json_encode(array("result" => 1, "errors" => $errors)));
-            } else {
-                $response = new Response(json_encode(array("result" => 0)));
-            }
+            $errors = $this->trans($errors);
+            $response = new Response(json_encode(array("result" => count($errors) > 0 ? 1 : 0, "errors" => $errors)));
         } catch (DAOUnsupportedOperationException $exc) {
-            $response = new Response(json_encode(array("result" => 2, "errors" => array($this->translator->trans('errors.table.column.conversion')))));
+            $response = new Response(json_encode(array("result" => 1, "errors" => array($this->translator->trans('errors.table.column.conversion')))));
         }
         $response->headers->set('Content-Type', 'application/json');
         return $response;
     }
 
     /**
-     * @Route("/DataTable/{table_id}/row/insert", name="DataTable_row_insert")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/row/insert", name="DataTable_row_insert", methods={"POST"})
+     * @param Request $request
      * @param $table_id
      * @return Response
      */
-    public function insertRowAction($table_id)
+    public function insertRowAction(Request $request, $table_id)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         $this->service->insertRow($table_id);
         $response = new Response(json_encode(array("result" => 0)));
         $response->headers->set('Content-Type', 'application/json');
@@ -301,8 +356,7 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{table_id}/row/{row_id}/update/{prefixed}", name="DataTable_row_update", defaults={"prefixed":0})
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/row/{row_id}/update/{prefixed}", name="DataTable_row_update", defaults={"prefixed":0}, methods={"POST"})
      * @param Request $request
      * @param $table_id
      * @param $row_id
@@ -311,6 +365,12 @@ class DataTableController extends AExportableTabController
      */
     public function updateRowAction(Request $request, $table_id, $row_id, $prefixed = 0)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         $this->service->updateRow($table_id, $row_id, $request->get("values"), $prefixed == 1);
         $response = new Response(json_encode(array("result" => 0)));
         $response->headers->set('Content-Type', 'application/json');
@@ -318,19 +378,28 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{object_ids}/export/{format}", name="DataTable_export", defaults={"format":"compressed"})
-     * @param $object_ids
+     * @Route("/DataTable/{instructions}/export/{format}", name="DataTable_export", defaults={"format":"yml"})
+     * @param string $instructions
      * @param string $format
      * @return Response
      */
-    public function exportAction($object_ids, $format = ExportService::FORMAT_COMPRESSED)
+    public function exportAction($instructions, $format = "yml")
     {
-        return parent::exportAction($object_ids, $format);
+        return parent::exportAction($instructions, $format);
     }
 
     /**
-     * @Route("/DataTable/import", name="DataTable_import")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{object_ids}/instructions/export", name="DataTable_export_instructions")
+     * @param $object_ids
+     * @return Response
+     */
+    public function exportInstructionsAction($object_ids)
+    {
+        return parent::exportInstructionsAction($object_ids);
+    }
+
+    /**
+     * @Route("/DataTable/import", name="DataTable_import", methods={"POST"})
      * @param Request $request
      * @return Response
      */
@@ -350,8 +419,7 @@ class DataTableController extends AExportableTabController
     }
 
     /**
-     * @Route("/DataTable/{table_id}/csv/{restructure}/{header}/{delimiter}/{enclosure}/import", name="DataTable_csv_import")
-     * @Method(methods={"POST"})
+     * @Route("/DataTable/{table_id}/csv/{restructure}/{header}/{delimiter}/{enclosure}/import", name="DataTable_csv_import", methods={"POST"})
      * @param Request $request
      * @param $table_id
      * @param $restructure
@@ -362,6 +430,12 @@ class DataTableController extends AExportableTabController
      */
     public function importCsvAction(Request $request, $table_id, $restructure, $header, $delimiter, $enclosure)
     {
+        if (!$this->service->canBeModified($table_id, $request->get("objectTimestamp"), $errorMessages)) {
+            $response = new Response(json_encode(array("result" => 1, "errors" => $this->trans($errorMessages))));
+            $response->headers->set('Content-Type', 'application/json');
+            return $response;
+        }
+
         try {
             $this->service->importFromCsv(
                 $table_id,
